Split positioning code into a separate source file.
